Address review questions
