+++ spring-integration-core/src/main/java/org/springframework/integration/transformer/support/ProtoHeaders.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 the original author or authors.
+ * Copyright 2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,7 +20,6 @@ package org.springframework.integration.transformer.support;
  * Pre-defined names and prefixes for Protocol Buffers related headers.
  *
  * @author Christian Tzolov
- *
  * @since 6.1
  */
 public final class ProtoHeaders {
