+++ spring-integration-core/src/test/resources/proto/ProtoTest.proto
